requirements.txt
Flask
Flask-Bootstrap
Flask-Login
Flask-SQLAlchemy
Flask-WTF
SQLAlchemy
pydeck
pandas
plotly
Flask-Migrate
python-dotenv
pymysql
gunicorn
gevent
scikit-learn
matplotlib
rq